rest-api-advanced: refactor and fixing bags for final review

+++ service/src/main/java/com/epam/esm/service/impl/GiftCertificateServiceImpl.java
@@ -1,22 +1,17 @@
 package com.epam.esm.service.impl;
 
-import com.epam.esm.hibernate.CertificateRepository;
 import com.epam.esm.converter.GiftCertificateDTOToEntityConverter;
 import com.epam.esm.converter.GiftCertificateEntityToDTOConverter;
 import com.epam.esm.dto.GiftCertificateDTO;
-import com.epam.esm.entity.GiftCertificate;
 import com.epam.esm.exception.DaoException;
 import com.epam.esm.exception.ServiceException;
+import com.epam.esm.hibernate.CertificateRepository;
 import com.epam.esm.persistence.GiftCertificateEntity;
 import com.epam.esm.service.GiftCertificateService;
 import com.epam.esm.validator.PreparedValidatorChain;
 import com.epam.esm.validator.realisation.IntermediateCertificateLink;
 import com.epam.esm.validator.realisation.IntermediateSortLink;
-import com.epam.esm.validator.realisation.certificate.CertificateNameValidatorLink;
-import com.epam.esm.validator.realisation.certificate.DescriptionValidatorLink;
-import com.epam.esm.validator.realisation.certificate.DurationValidatorLink;
-import com.epam.esm.validator.realisation.certificate.PriceValidatorLink;
-import com.epam.esm.validator.realisation.certificate.TagsValidatorLink;
+import com.epam.esm.validator.realisation.certificate.*;
 import com.epam.esm.validator.realisation.sort.FieldValidatorLink;
 import com.epam.esm.validator.realisation.sort.MethodValidatorLink;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,9 +23,9 @@ import java.util.stream.Collectors;
 @Service
 public class GiftCertificateServiceImpl implements GiftCertificateService {
 
-    private GiftCertificateEntityToDTOConverter converterToDto;
-    private GiftCertificateDTOToEntityConverter converterToEntity;
-    private CertificateRepository repository;
+    private final GiftCertificateEntityToDTOConverter converterToDto;
+    private final GiftCertificateDTOToEntityConverter converterToEntity;
+    private final CertificateRepository repository;
 
     @Autowired
     public GiftCertificateServiceImpl(GiftCertificateEntityToDTOConverter converterToDto,
@@ -41,16 +36,12 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
         this.repository = repository;
     }
 
-    public GiftCertificateServiceImpl() {
-    }
-
     @Override
     public List<GiftCertificateDTO> findAll(int limit, int page) throws ServiceException {
         try {
-            List<GiftCertificateEntity> listFromDao = repository.findAll(limit, page);
-            return createResultList(listFromDao);
+            return createResultList(repository.findAll(limit, page));
         } catch (DaoException e) {
-            throw new ServiceException("exception in dao", e.getCause());
+            throw new ServiceException(e.getCause());
         }
     }
 
@@ -60,7 +51,7 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
             Optional<GiftCertificateEntity> certificateFromDao = repository.find(id);
             return certificateFromDao.map(converterToDto);
         } catch (DaoException e) {
-            throw new ServiceException("exception in dao", e.getCause());
+            throw new ServiceException(e.getCause());
         }
 
     }
@@ -78,7 +69,7 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
             try {
                 return repository.create(certificate);
             } catch (DaoException e) {
-                throw new ServiceException("exception in dao", e.getCause());
+                throw new ServiceException(e.getCause());
             }
         }
         return 0;
@@ -93,31 +84,66 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
         if (chain.validate(certificate)) {
             GiftCertificateEntity certificateForUpdate = converterToEntity.apply((certificate));
             try {
-                repository.update(certificateForUpdate, id);
-                return true;
+                Optional<GiftCertificateEntity> certFromDao = repository.find(id);
+                if (certFromDao.isPresent()) {
+                    insertDataForUpdate(certificateForUpdate, certFromDao.get());
+                    repository.update(certificateForUpdate);
+                    return true;
+                } else {
+                    return false;
+                }
             } catch (DaoException e) {
-                throw new ServiceException("exception in dao", e.getCause());
+                throw new ServiceException(e.getCause());
             }
         }
         return false;
     }
 
+    private void insertDataForUpdate
+            (GiftCertificateEntity example, GiftCertificateEntity cert) {
+        if (example.getName() != null) {
+            cert.setName(example.getName());
+        }
+        if (example.getDescription() != null) {
+            cert.setDescription(example.getDescription());
+        }
+        if (example.getLastUpdateDate() != null) {
+            cert.setLastUpdateDate(example.getLastUpdateDate());
+        }
+        if (example.getCreateDate() != null) {
+            cert.setCreateDate(example.getCreateDate());
+        }
+        if (example.getDuration() != 0) {
+            cert.setDuration(example.getDuration());
+        }
+        if (example.getPrice() != 0) {
+            cert.setPrice(example.getPrice());
+        }
+        if (example.getTagsDependsOnCertificate() != null) {
+            cert.setTagsDependsOnCertificate(example.getTagsDependsOnCertificate());
+        }
+    }
+
     @Override
     public void delete(long id) throws ServiceException {
         try {
-            repository.delete(id);
+            Optional<GiftCertificateEntity> tagWrapper = repository.find(id);
+            if (tagWrapper.isPresent()) {
+                repository.delete(id);
+            } else {
+                throw new ServiceException("not found");
+            }
         } catch (DaoException e) {
-            throw new ServiceException("exception in dao", e.getCause());
+            throw new ServiceException(e.getCause());
         }
     }
 
     @Override
     public List<GiftCertificateDTO> findByPartOfName(String partOfName, int limit, int page) throws ServiceException {
         try {
-            List<GiftCertificateEntity> listFromDao = repository.searchByName(partOfName, limit, page);
-            return createResultList(listFromDao);
+            return createResultList(repository.searchByName(partOfName, limit, page));
         } catch (DaoException e) {
-            throw new ServiceException("exception in dao", e.getCause());
+            throw new ServiceException(e.getCause());
         }
     }
 
@@ -125,20 +151,18 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
     public List<GiftCertificateDTO> findByPartOfDescription(String partOfDescription, int limit, int page)
             throws ServiceException {
         try {
-            List<GiftCertificateEntity> listFromDao = repository.searchByDescription(partOfDescription, limit, page);
-            return createResultList(listFromDao);
+            return createResultList(repository.searchByDescription(partOfDescription, limit, page));
         } catch (DaoException e) {
-            throw new ServiceException("exception in dao", e.getCause());
+            throw new ServiceException(e.getCause());
         }
     }
 
     @Override
     public List<GiftCertificateDTO> findByTag(String nameOfTag, int limit, int page) throws ServiceException {
         try {
-            List<GiftCertificateEntity> listFromDao = repository.searchByTag(nameOfTag, limit, page);
-            return createResultList(listFromDao);
+            return createResultList(repository.searchByTag(nameOfTag, limit, page));
         } catch (DaoException e) {
-            throw new ServiceException("exception in dao", e.getCause());
+            throw new ServiceException(e.getCause());
         }
     }
 
@@ -153,7 +177,7 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
             chain.linkWith(new FieldValidatorLink()).linkWith(new MethodValidatorLink());
             if (chain.validate(tempMapForValidateParams)) {
                 List<GiftCertificateEntity> listFromDao;
-                if(field.equals("name_of_certificate")) {
+                if (field.equals("name_of_certificate")) {
                     listFromDao = repository.sortCertificatesByName(method, limit, page);
                 } else {
                     listFromDao = repository.sortCertificatesByCreateDate
@@ -163,7 +187,7 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
             }
             return new ArrayList<>();
         } catch (DaoException e) {
-            throw new ServiceException("exception in dao", e.getCause());
+            throw new ServiceException(e.getCause());
         }
     }
 
@@ -172,6 +196,4 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
                 .map(certificate -> converterToDto.apply(certificate))
                 .collect(Collectors.toList());
     }
-
-
 }
